setup and clear next edit window manager on optInContinueFeature, pass context to vscode messenger
